Assess decoy rate in data

This function assesses the number of quantifications (typically peptides) that are decoys (false-positive) versus true identifications.

assess_decoy_rate(data, column = "FullPeptideName", column_decoy = "decoy")

Arguments

data

A data frame that contains at least a column named "FullPeptideName" and "decoy".

column

The column name of the Peptide identifier. Default: FullPeptideName.

column_decoy

The column name of the decoy column. Default: decoy.

Value

Message detailing the number of decoys, non-decoys, and the ratio.

Details

A printout is generated to indicate the number of non-decoy, decoy peptides and the rate of decoy vs non-decoy peptides. Unique peptides are counted, so a precursor with different charge states is counted as one peptide. In the column "decoy" the values need to be 1,0 or TRUE and FALSE.

Author

Peter Blattmann

Examples

 data("OpenSWATH_data", package="SWATH2stats")
 data <- OpenSWATH_data
 assess_decoy_rate(data)
#> Number of non-decoy identifiers: 273
#> Number of decoy identifiers: 11
#> Decoy rate: 0.0403