Get data ready for use by MSstats.

Though SWATH2stats uses very similar format as MSstats, some coercion is required to convert the data into the format for MSstats.

convert4MSstats(
  data,
  replace_values = TRUE,
  replace_colnames = TRUE,
  replace_unimod = TRUE
)

Arguments

data

A data frame containing SWATH data.

replace_values

Option to indicate if negative and 0 values should be replaced with NA.

replace_colnames

Option to indicate if column names should be renamed and columns reduced to the necessary columns for MSstats.

replace_unimod

Option to indicate if Unimod Identifier should be replaced from ":" to "_".

Value

Returns a data frame in the appropriate format for MSstats.

Details

This functions selects the columns necessary for MSstats and renames them if necessary.

The necessary columns are selected and three columns renamed: FullPeptideName -> PeptideSequence Charge -> PrecursorCharge filename -> File

References

Choi M, Chang CY, Clough T, Broudy D, Killeen T, MacLean B, Vitek O. MSstats: an R package for statistical analysis of quantitative mass spectrometry-based proteomic experiments.Bioinformatics. 2014 Sep 1;30(17):2524-6. doi: 10.1093/bioinformatics/btu305.

Author

Peter Blattmann

Examples

 data("OpenSWATH_data", package="SWATH2stats")
 data("Study_design", package="SWATH2stats")
 data <- sample_annotation(OpenSWATH_data, Study_design)
 data.filtered.decoy <- filter_mscore(data, 0.01)
#> Dimension difference: 931, 0
 raw <- disaggregate(data.filtered.decoy)
#> The library contains 6 transitions per precursor.
#>                   
#> The data table was transformed into a table containing one row per transition.
 data.mapDIA <- convert4MSstats(raw)
#> One or several columns required by MSstats were not in the data. 
#>                 The columns were created and filled with NAs.
#> Missing columns: ProductCharge, IsotopeLabelType
#> IsotopeLabelType was filled with light.
#> Warning: Intensity values that were 0, were replaced by NA